feat: 당첨금 합 타입 변경(int -> long)

+++ src/main/java/domain/StatisticsService.java
@@ -27,10 +27,10 @@ public class StatisticsService {
     }
 
     public Profit calculateProfit(WinningStatistics winningStatistics, int lottoTicketNumber) {
-        int sum = 0;
+        long sum = 0;
         Map<LottoPrize, Integer> prizeCounter = winningStatistics.getPrizeCounter();
         for (LottoPrize lottoPrize : LottoPrize.values()) {
-            sum += prizeCounter.get(lottoPrize) * lottoPrize.getMoney();
+            sum += (long) prizeCounter.get(lottoPrize) * lottoPrize.getMoney();
         }
         return new Profit((double) sum / (lottoTicketNumber * LottoTicket.LOTTO_PRICE));
     }
